Refactor of model.py into a model module where the tables are broken into their respective categories. Some classes are in their own individual files to avoid circular imports.
