+++ hydra_base/db/model/scenario/__init__.py
@@ -1,0 +1,2 @@
+from .resourcegroupitem import *
+from .scenario import *
